[WFLY-5055] Remove the JobXmlResolverService as it's needed during deployment processing. Using a services started in a DUP and consumed in a DUP is bad practice.
