+++ batch/extension-jberet/src/main/java/org/wildfly/extension/batch/jberet/deployment/BatchEnvironmentProcessor.java
@@ -22,19 +22,14 @@
 
 package org.wildfly.extension.batch.jberet.deployment;
 
+import java.util.concurrent.ExecutorService;
 import javax.enterprise.inject.spi.BeanManager;
 import javax.transaction.TransactionManager;
-import java.io.IOException;
-import java.util.concurrent.ExecutorService;
 
 import org.jberet.repository.JobRepository;
 import org.jberet.spi.BatchEnvironment;
-import org.jberet.spi.JobXmlResolver;
 import org.jboss.as.controller.capability.CapabilityServiceSupport;
-import org.jboss.as.ee.structure.DeploymentType;
-import org.jboss.as.ee.structure.DeploymentTypeMarker;
 import org.jboss.as.ee.weld.WeldDeploymentMarker;
-import org.jboss.as.server.Services;
 import org.jboss.as.server.deployment.Attachments;
 import org.jboss.as.server.deployment.DeploymentPhaseContext;
 import org.jboss.as.server.deployment.DeploymentUnit;
@@ -46,8 +41,6 @@ import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceName;
 import org.jboss.msc.service.ServiceTarget;
 import org.jboss.msc.value.ImmediateValue;
-import org.jboss.vfs.VirtualFile;
-import org.jboss.vfs.VirtualFileFilter;
 import org.wildfly.extension.batch.jberet.BatchServiceNames;
 import org.wildfly.extension.batch.jberet._private.BatchLogger;
 import org.wildfly.extension.batch.jberet._private.Capabilities;
@@ -96,7 +89,7 @@ public class BatchEnvironmentProcessor implements DeploymentUnitProcessor {
             final CapabilityServiceSupport support = deploymentUnit.getAttachment(Attachments.CAPABILITY_SERVICE_SUPPORT);
 
             // Create the batch environment
-            final BatchEnvironmentService service = new BatchEnvironmentService(moduleClassLoader, deploymentUnit.getName());
+            final BatchEnvironmentService service = new BatchEnvironmentService(moduleClassLoader, WildFlyJobXmlResolver.of(moduleClassLoader, deploymentUnit), deploymentUnit.getName());
             final ServiceBuilder<BatchEnvironment> serviceBuilder = serviceTarget.addService(BatchServiceNames.batchEnvironmentServiceName(deploymentUnit), service);
 
             // Add a dependency to the thread-pool
@@ -117,34 +110,6 @@ public class BatchEnvironmentProcessor implements DeploymentUnitProcessor {
                 serviceBuilder.addDependency(BatchServiceNames.beanManagerServiceName(deploymentUnit), BeanManager.class, service.getBeanManagerInjector());
             }
 
-            // Get the root file
-            final VirtualFile root = deploymentUnit.getAttachment(Attachments.DEPLOYMENT_ROOT).getRoot();
-            VirtualFile jobsDir;
-            // Only files in the META-INF/batch-jobs directory
-            if (DeploymentTypeMarker.isType(DeploymentType.WAR, deploymentUnit)) {
-                jobsDir = root.getChild("WEB-INF/classes/META-INF/batch-jobs");
-            } else {
-                jobsDir = root.getChild("META-INF/batch-jobs");
-            }
-            final JobXmlResolverService jobXmlResolverService;
-            if (jobsDir != null && jobsDir.exists()) {
-                try {
-                    // Create the job XML resolver service with the files allowed to be used
-                    jobXmlResolverService = new JobXmlResolverService(moduleClassLoader, jobsDir.getChildren(JobXmlFilter.INSTANCE));
-                } catch (IOException e) {
-                    throw BatchLogger.LOGGER.errorProcessingBatchJobsDir(e);
-                }
-            } else {
-                // This is likely not a batch deployment, creates a no-op service
-                jobXmlResolverService = new JobXmlResolverService();
-            }
-            // Install the job XML resolver service
-            final ServiceBuilder<JobXmlResolver> jobXmlServiceBuilder = Services.addServerExecutorDependency(
-                    serviceTarget.addService(BatchServiceNames.jobXmlResolverServiceName(deploymentUnit), jobXmlResolverService),
-                    jobXmlResolverService.getExecutorServiceInjector(), false);
-            // Add a dependency to the job XML resolver service
-            serviceBuilder.addDependency(BatchServiceNames.jobXmlResolverServiceName(deploymentUnit), JobXmlResolver.class, service.getJobXmlResolverInjector());
-
             // No deployment defined repository, use the default
             if (jobRepository == null) {
                 final ServiceName defaultJobRepository = support.getCapabilityServiceName(Capabilities.DEFAULT_JOB_REPOSITORY_CAPABILITY.getName());
@@ -157,22 +122,11 @@ public class BatchEnvironmentProcessor implements DeploymentUnitProcessor {
                 serviceBuilder.addDependency(RequestController.SERVICE_NAME, RequestController.class, service.getRequestControllerInjector());
             }
 
-            jobXmlServiceBuilder.install();
             serviceBuilder.install();
         }
     }
 
     @Override
     public void undeploy(DeploymentUnit context) {
-    }
-
-    private static class JobXmlFilter implements VirtualFileFilter {
-
-        static final JobXmlFilter INSTANCE = new JobXmlFilter();
-
-        @Override
-        public boolean accepts(final VirtualFile file) {
-            return file.isFile() && file.getName().endsWith(".xml");
-        }
     }
 }
