+++ batch/extension/src/main/java/org/wildfly/extension/batch/deployment/BatchEnvironmentProcessor.java
@@ -22,18 +22,13 @@
 
 package org.wildfly.extension.batch.deployment;
 
-import java.io.IOException;
 import java.util.concurrent.ExecutorService;
 import javax.enterprise.inject.spi.BeanManager;
 import javax.transaction.TransactionManager;
 
 import org.jberet.spi.BatchEnvironment;
-import org.jberet.spi.JobXmlResolver;
 import org.jboss.as.ee.component.EEModuleDescription;
-import org.jboss.as.ee.structure.DeploymentType;
-import org.jboss.as.ee.structure.DeploymentTypeMarker;
 import org.jboss.as.ee.weld.WeldDeploymentMarker;
-import org.jboss.as.server.Services;
 import org.jboss.as.server.deployment.Attachments;
 import org.jboss.as.server.deployment.DeploymentPhaseContext;
 import org.jboss.as.server.deployment.DeploymentUnit;
@@ -44,11 +39,9 @@ import org.jboss.modules.Module;
 import org.jboss.msc.service.ServiceBuilder;
 import org.jboss.msc.service.ServiceTarget;
 import org.jboss.msc.value.ImmediateValue;
-import org.jboss.vfs.VirtualFile;
-import org.jboss.vfs.VirtualFileFilter;
 import org.wildfly.extension.batch.BatchServiceNames;
 import org.wildfly.extension.batch._private.BatchLogger;
-import org.wildfly.extension.batch.jberet.deployment.JobXmlResolverService;
+import org.wildfly.extension.batch.jberet.deployment.WildFlyJobXmlResolver;
 import org.wildfly.extension.batch.job.repository.JobRepositoryFactory;
 import org.wildfly.extension.requestcontroller.RequestController;
 import org.wildfly.jberet.services.BatchEnvironmentService;
@@ -78,7 +71,7 @@ public class BatchEnvironmentProcessor implements DeploymentUnitProcessor {
             final EEModuleDescription moduleDescription = deploymentUnit.getAttachment(org.jboss.as.ee.component.Attachments.EE_MODULE_DESCRIPTION);
 
             // Create the batch environment
-            final BatchEnvironmentService service = new BatchEnvironmentService(moduleClassLoader, deploymentUnit.getName());
+            final BatchEnvironmentService service = new BatchEnvironmentService(moduleClassLoader, WildFlyJobXmlResolver.of(moduleClassLoader, deploymentUnit), deploymentUnit.getName());
             // Set the value for the job-repository, this can't be a capability as the JDBC job repository cannot be constructed
             // until deployment time because the default JNDI data-source name is only known during DUP processing
             service.getJobRepositoryInjector().setValue(new ImmediateValue<>(JobRepositoryFactory.getInstance().getJobRepository(moduleDescription)));
@@ -93,35 +86,6 @@ public class BatchEnvironmentProcessor implements DeploymentUnitProcessor {
                 serviceBuilder.addDependency(BatchServiceNames.beanManagerServiceName(deploymentUnit), BeanManager.class, service.getBeanManagerInjector());
             }
 
-            // Get the root file
-            final VirtualFile root = deploymentUnit.getAttachment(Attachments.DEPLOYMENT_ROOT).getRoot();
-            VirtualFile jobsDir = null;
-            // Only files in the META-INF/batch-jobs directory
-            if (DeploymentTypeMarker.isType(DeploymentType.WAR, deploymentUnit)) {
-                jobsDir = root.getChild("WEB-INF/classes/META-INF/batch-jobs");
-            } else {
-                jobsDir = root.getChild("META-INF/batch-jobs");
-            }
-            final JobXmlResolverService jobXmlResolverService;
-            if (jobsDir != null && jobsDir.exists()) {
-                try {
-                    // Create the job XML resolver service with the files allowed to be used
-                    jobXmlResolverService = new JobXmlResolverService(moduleClassLoader, jobsDir.getChildren(JobXmlFilter.INSTANCE));
-                } catch (IOException e) {
-                    throw BatchLogger.LOGGER.errorProcessingBatchJobsDir(e);
-                }
-            } else {
-                // This is likely not a batch deployment, creates a no-op service
-                jobXmlResolverService = new JobXmlResolverService();
-            }
-            // Install the job XML resolver service
-            Services.addServerExecutorDependency(
-                    serviceTarget.addService(org.wildfly.extension.batch.jberet.BatchServiceNames.jobXmlResolverServiceName(deploymentUnit), jobXmlResolverService),
-                    jobXmlResolverService.getExecutorServiceInjector(), false)
-                    .install();
-            // Add a dependency to the job XML resolver service
-            serviceBuilder.addDependency(BatchServiceNames.jobXmlResolverServiceName(deploymentUnit), JobXmlResolver.class, service.getJobXmlResolverInjector());
-
             if (rcPresent) {
                 serviceBuilder.addDependency(RequestController.SERVICE_NAME, RequestController.class, service.getRequestControllerInjector());
             }
@@ -132,15 +96,5 @@ public class BatchEnvironmentProcessor implements DeploymentUnitProcessor {
 
     @Override
     public void undeploy(DeploymentUnit context) {
-    }
-
-    private static class JobXmlFilter implements VirtualFileFilter {
-
-        static final JobXmlFilter INSTANCE = new JobXmlFilter();
-
-        @Override
-        public boolean accepts(final VirtualFile file) {
-            return file.isFile() && file.getName().endsWith(".xml");
-        }
     }
 }
